Workaround for GCS backend bug
We've seen in production that the v1 persister occasionally hangs in an
infinite retry loop of 503s from cloud storage. The root cause is that
the GCS backend actually isn't accepting earlier parts of a resumable
upload, so when the later offsets are sent, the backend loops on
rejecting it.

LiveRamp has an enterprise support ticket with Google to work on a
permanent solution, but this workaround should help in the mean time, as
retried requests will send the entire fragment contents instead of
chunking it.